+++ api/src/main/java/org/snomed/snap2snomed/service/MapViewService.java
@@ -295,9 +295,9 @@ public class MapViewService {
     return mapRepository.findSourceByMapId(mapId).get().getAdditionalColumnsMetadata();
   }
 
-  public String[] getExportHeader(Long mapId) {
-
-    ArrayList<String> exportHeader = new ArrayList<String>(Arrays.asList("\uFEFF" + "Source code", "Source display"));
+  public String[] getExportHeader(Long mapId, List<String> extraColumns) {
+    
+    ArrayList<String> exportHeader = new ArrayList<String>(Arrays.asList("\ufeff" + "Source code", "Source display"));
 
     final List<AdditionalCodeColumn> additionalCodeColumnList = this.getAdditionalColumnsMetadata(mapId);
     if (additionalCodeColumnList != null && additionalCodeColumnList.size() > 0) {
@@ -306,6 +306,28 @@ public class MapViewService {
       }
     }
     exportHeader.addAll(Arrays.asList("Target code", "Target display", "Relationship type code", "Relationship type display", "No map flag", "Status"));
+
+    if (extraColumns != null) {
+      for (String extraColumn : extraColumns) {
+        switch(extraColumn.toUpperCase()) {
+          case "NOTES":
+            exportHeader.add("Notes");
+            break;
+          case "ASSIGNEDAUTHOR":
+            exportHeader.add("Assigned author");
+            break;
+          case "ASSIGNEDREVIEWER":
+            exportHeader.add("Assigned reviewer");
+            break;
+          case "LASTAUTHOR":
+            exportHeader.add("Last author");
+            break;
+          case "LASTREVIEWER":
+            exportHeader.add("Last reviewer");
+            break;
+        }
+      }
+    }
 
     return exportHeader.toArray(new String[0]);